+++ crates/net/network/tests/it/main.rs
@@ -2,6 +2,5 @@ mod connect;
 mod requests;
 mod session;
 mod startup;
-mod transaction;
 
 fn main() {}
